+++ lib/flutter_pop_kit.dart
@@ -1,3 +1,7 @@
 // Alert Dialog
 export 'sources/flutter_pop_alert_dialog.dart';
 export 'functions/flutter_pop_alert_dialog.dart';
+
+// Bottom Sheet
+export 'sources/flutter_pop_bottom_sheet.dart';
+export 'functions/flutter_pop_bottom_sheet.dart';